+++ index.html
@@ -8,6 +8,6 @@
   </head>
   <body>
     <div id="app"></div>
-    <script type="module" src="/src/test.tsx"></script>
+    <script type="module" src="/src/App.tsx"></script>
   </body>
 </html>
