+++ package.json
@@ -16,6 +16,7 @@
     "@babel/preset-typescript": "^7.26.0",
     "@types/react": "^19.0.3",
     "@types/react-dom": "^19.0.2",
+    "babel-cli": "^6.26.0",
     "babel-loader": "^9.2.1",
     "typescript": "~5.6.2",
     "vite": "^6.0.5",
