+++ src/components/Content.tsx
@@ -1,0 +1,11 @@
+import { createElement } from "../../libs/jsx-runtime";
+
+function Content() {
+  return (
+    <main>
+      <p>This is a sample content.</p>
+    </main>
+  );
+}
+
+export default Content;
