+++ src/components/Header.tsx
@@ -1,0 +1,11 @@
+import { createElement } from "../../libs/jsx-runtime";
+
+function Header() {
+  return (
+    <header>
+      <h1>My React Clone</h1>
+    </header>
+  );
+}
+
+export default Header;
